dxt-packages/crawl4ai-dxt-correct/requirements.txt
crawl4ai>=0.3.0
fastmcp>=0.1.0
pydantic>=2.0.0
asyncio
typing-extensions
python-dotenv>=1.0.0

# File processing - explicit dependencies to avoid conflicts
markitdown>=0.0.1a2
pdfminer-six>=20250506
mammoth>=1.9.1
openpyxl>=3.1.5
python-pptx>=1.0.2
xlrd>=2.0.1

googlesearch-python>=1.3.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0

# YouTube transcript extraction (youtube-transcript-api v1.1.0+)
youtube-transcript-api>=1.1.0
